Move createElement function to logic.js
